Move manifest machinery into class
- Don't keep sub manifests once finished processing each piece
- use the payload (root) CID for the piece filename
