+++ src/cmd/pack.ts
@@ -5,6 +5,7 @@ import {
   existsSync,
   mkdirSync,
   readFileSync,
+  renameSync,
   writeFile,
 } from "node:fs";
 import { open } from "node:fs/promises";
@@ -16,14 +17,8 @@ import { Block } from "@ipld/unixfs";
 import { CAREncoderStream } from "ipfs-car";
 import { CID } from "multiformats/cid";
 
-import { mergeDeep } from "../contents.js";
 import { iterateFilesFromPathsWithSize } from "../files.js";
-import type {
-  SubManifest,
-  SuperManifest,
-  SuperManifestContentEntry,
-  UserMetadata,
-} from "../manifest.d.js";
+import { Manifest, type UserMetadata } from "../manifest.js";
 import parseBytes from "../parseBytes.js";
 import { createDirectoryEncoderStream } from "../unixfs.js";
 
@@ -31,9 +26,6 @@ import { createDirectoryEncoderStream } from "../unixfs.js";
 const placeholderCID = CID.parse(
   "bafybeigdyrzt5sfp7udm7hu76uh7y26nf3efuylqabf3oclgtqy55fbzdi"
 );
-
-const currentSpecURL =
-  "https://raw.githubusercontent.com/fidlabs/data-prep-standard/refs/heads/main/specification/v0/FilecoinDataPreparationManifestSpecification.md";
 
 export default async function pack(
   filePaths: string[],
@@ -49,45 +41,38 @@ export default async function pack(
   const targetCarSize = parseBytes(opts.targetCarSize);
   console.log("pack", filePaths, opts, targetCarSize);
 
-  if (opts.specVersion !== "0.1.0") {
-    throw new Error(
-      `Unsupported schema version: ${opts.specVersion}. Only 0.1.0 is supported.`
-    );
-  }
-
   // Check if output directory exists, if not create it
   if (!existsSync(opts.output)) {
     mkdirSync(opts.output, { recursive: true });
   }
 
-  const metadata = JSON.parse(
-    readFileSync(opts.metadata, "utf-8")
-  ) as UserMetadata;
-
-  const uuid = crypto.randomUUID();
-
-  // Create one or more streams of files to pack
-  const subManifests: { rootCID: CID; subManifest: SubManifest }[] = [];
-  let nPiece = 0;
+  const manifest = new Manifest(
+    JSON.parse(readFileSync(opts.metadata, "utf-8")) as UserMetadata,
+    opts.specVersion,
+    { lite: opts.lite ?? undefined }
+  );
 
   for await (const files of iterateFilesFromPathsWithSize(
     filePaths,
     targetCarSize
   )) {
     let rootCID = placeholderCID;
+    let pieceCID = placeholderCID;
+
+    const subManifest = manifest.newSubManifest();
+    // We use a UUID for a temporary piece file name, we will renamr it later
+    // after the root (payload) CID is known.
+    const pieceTemporaryFilename = `piece-${crypto.randomUUID()}.car`;
 
     console.log(
-      `Piece ${String(nPiece)}: files to pack: `,
+      `New piece: files to pack: `,
       files.map((f) => f.name)
     );
 
-    const { stream, close } = createDirectoryEncoderStream(files, {
-      metadata,
-      uuid,
-      specVersion: opts.specVersion,
-      lite: opts.lite ?? false,
-      spec: currentSpecURL,
-    });
+    const { stream, subManifestPromise } = createDirectoryEncoderStream(
+      files,
+      subManifest
+    );
 
     await stream
       .pipeThrough(
@@ -95,6 +80,9 @@ export default async function pack(
           transform(block, controller) {
             // we visit every block, the last one will be the root block
             rootCID = CID.parse(block.cid.toString());
+
+            // TODO: work out the proper piece CID (CommP)
+            pieceCID = rootCID;
             // console.log("Root CID:", rootCID.toString());
             controller.enqueue(block);
           },
@@ -103,58 +91,32 @@ export default async function pack(
       .pipeThrough(new CAREncoderStream([placeholderCID]))
       .pipeTo(
         Writable.toWeb(
-          createWriteStream(
-            join(
-              opts.output,
-              `piece-${nPiece.toString().padStart(4, "0")}.car`
-            ),
-            { autoClose: true }
-          )
+          createWriteStream(join(opts.output, pieceTemporaryFilename), {
+            autoClose: true,
+          })
         )
       );
 
     // console.log("Packing completed, root CID:", rootCID.toString());
 
-    const subManifest = await close;
+    await subManifestPromise;
 
     // update roots in CAR header
-    const fd = await open(
-      join(opts.output, `piece-${nPiece.toString().padStart(4, "0")}.car`),
-      "r+"
-    );
+    const fd = await open(join(opts.output, pieceTemporaryFilename), "r+");
     await CarWriter.updateRootsInFile(fd, [rootCID]);
     await fd.close();
 
-    subManifests.push({ rootCID, subManifest });
-    nPiece++;
-  }
+    renameSync(
+      join(opts.output, pieceTemporaryFilename),
+      join(opts.output, `piece-${rootCID.toString()}.car`)
+    );
 
-  const superManifest: SuperManifest = {
-    "@spec": currentSpecURL,
-    "@spec_version": opts.specVersion,
-    ...metadata,
-    uuid,
-    n_pieces: subManifests.length,
-    pieces: subManifests.map(({ rootCID }) => {
-      return {
-        piece_cid: "TODO: calculate commP",
-        payload_cid: rootCID.toString(),
-      };
-    }),
-  };
-
-  if (!opts.lite) {
-    const contents: SuperManifestContentEntry[] = [];
-    // TODO: should be pieceCID
-    subManifests.forEach(({ rootCID, subManifest }) => {
-      mergeDeep(contents, rootCID, ...(subManifest.contents ?? []));
-    });
-    superManifest.contents = contents;
+    manifest.addPiece(subManifest, pieceCID, rootCID);
   }
 
   writeFile(
     join(opts.output, "manifest.json"),
-    JSON.stringify(superManifest, null, 2),
+    JSON.stringify(manifest, null, 2),
     (error) => {
       if (error) throw error;
     }
